extracting the field config registry factory method into its own module.
