+++ packages/grafana-data/src/index.ts
@@ -19,3 +19,4 @@ export {
   RangeValueMatcherOptions,
 } from './transformations/matchers/valueMatchers/types';
 export { PanelPlugin } from './panel/PanelPlugin';
+export { createFieldConfigRegistry } from './panel/registryFactories';
