+++ packages/grafana-data/src/panel/PanelPlugin.ts
@@ -13,7 +13,8 @@ import { FieldConfigEditorBuilder, PanelOptionsEditorBuilder } from '../utils/Op
 import { ComponentClass, ComponentType } from 'react';
 import set from 'lodash/set';
 import { deprecationWarning } from '../utils';
-import { FieldConfigOptionsRegistry, standardFieldConfigEditorRegistry } from '../field';
+import { FieldConfigOptionsRegistry } from '../field';
+import { createFieldConfigRegistry } from './registryFactories';
 
 type StandardOptionConfig = {
   defaultValue?: any;
@@ -312,55 +313,7 @@ export class PanelPlugin<TOptions = any, TFieldConfigOptions extends object = an
    */
   useFieldConfig(config: SetFieldConfigOptionsArgs<TFieldConfigOptions> = {}) {
     // builder is applied lazily when custom field configs are accessed
-    this._initConfigRegistry = () => {
-      const registry = new FieldConfigOptionsRegistry();
-
-      // Add custom options
-      if (config.useCustomConfig) {
-        const builder = new FieldConfigEditorBuilder<TFieldConfigOptions>();
-        config.useCustomConfig(builder);
-
-        for (const customProp of builder.getRegistry().list()) {
-          customProp.isCustom = true;
-          customProp.category = [`${this.meta.name} options`].concat(customProp.category || []);
-          // need to do something to make the custom items not conflict with standard ones
-          // problem is id (registry index) is used as property path
-          // so sort of need a property path on the FieldPropertyEditorItem
-          customProp.id = 'custom.' + customProp.id;
-          registry.register(customProp);
-        }
-      }
-
-      for (let fieldConfigProp of standardFieldConfigEditorRegistry.list()) {
-        if (config.disableStandardOptions) {
-          const isDisabled = config.disableStandardOptions.indexOf(fieldConfigProp.id as FieldConfigProperty) > -1;
-          if (isDisabled) {
-            continue;
-          }
-        }
-        if (config.standardOptions) {
-          const customDefault: any = config.standardOptions[fieldConfigProp.id as FieldConfigProperty]?.defaultValue;
-          const customSettings: any = config.standardOptions[fieldConfigProp.id as FieldConfigProperty]?.settings;
-          if (customDefault) {
-            fieldConfigProp = {
-              ...fieldConfigProp,
-              defaultValue: customDefault,
-            };
-          }
-
-          if (customSettings) {
-            fieldConfigProp = {
-              ...fieldConfigProp,
-              settings: fieldConfigProp.settings ? { ...fieldConfigProp.settings, ...customSettings } : customSettings,
-            };
-          }
-        }
-
-        registry.register(fieldConfigProp);
-      }
-
-      return registry;
-    };
+    this._initConfigRegistry = () => createFieldConfigRegistry(config, this.meta.name);
 
     return this;
   }
